+++ src/test/java/com/github/jlangch/venice/impl/functions/CoreFunctionsTest.java
@@ -4598,9 +4598,21 @@ public class CoreFunctionsTest {
     public void test_partition_by() {
         final Venice venice = new Venice();
 
+        assertEquals("((1) (2) (3) (4) (5) (6))", venice.eval("(str (partition-by even? [1 2 3 4 5 6]))"));
         assertEquals("((1) (2 4) (3 5) (6))", venice.eval("(str (partition-by even? [1 2 4 3 5 6]))"));
         assertEquals("((A) (B B) (A))", venice.eval("(str (partition-by identity (seq \"ABBA\")))"));
         assertEquals("((1 1 1 1) (2 2) (3))", venice.eval("(str (partition-by identity [1 1 1 1 2 2 3]))"));
+    }
+
+    @Test
+    public void test_partition_at() {
+        final Venice venice = new Venice();
+
+        assertEquals("((1) (2) (3) (4) (5) (6))", venice.eval("(str (partition-at (fn [x] true) [1 2 3 4 5 6]))"));
+        assertEquals("((1 2 3 4 5 6))", venice.eval("(str (partition-at (fn [x] false) [1 2 3 4 5 6]))"));
+        assertEquals("((1) (2 3) (4 5) (6))", venice.eval("(str (partition-at even? [1 2 3 4 5 6]))"));
+        assertEquals("((1) (2) (4 3 5) (6))", venice.eval("(str (partition-at even? [1 2 4 3 5 6]))"));
+        assertEquals("((A B B A))", venice.eval("(str (partition-at identity (seq \"ABBA\")))"));
     }
 
     @Test